scripts/kconfig/symbol.c: fix misleading indentation
In scripts/kconfig/symbol.c at line 1119, there was a warning generated
by the compiler regarding misleading indentation. Specifically, the
'fprintf' statement was incorrectly aligned and appeared to be part of
the previous 'if' statement, when it was not.

To address this issue and improve code readability, we have properly
indented the 'fprintf' statement to ensure it is clearly separate from
the preceding 'if' statement.

Change-Id: If4205dbfecc3c49e953c9f9168a624e3b0095283
Signed-off-by: Kai Jones <19639817+488315@users.noreply.github.com>