Makefile: Set ld.lld as the Default Linker
This commit addresses a linker error when using clang-14, where the system is unable to find the "ld" linker executable. To resolve this issue, ld.lld is now set as the default linker. This change ensures that the linker command is executed successfully, preventing the error and allowing the project to build correctly.

Change-Id: I3b54fdc2c95569316b25687031154589818e2fe0
Signed-off-by: Kai Jones <19639817+488315@users.noreply.github.com>